Restore cdGunSight shader microcode

+++ src/render/cdGunSightShader.h
@@ -67,7 +67,7 @@ namespace cdGunSightFullbrightPixel {
 // Externs
 // ============================================================================
 extern void nglDxRegisterVShader(unsigned int* VS, const unsigned int* Microcode);  // ngl_dx_shader.o
-extern void nglDxRegisterPShader(unsigned int** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
+extern void nglDxRegisterPShader(unsigned long** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
 
 extern cdGunSightShader* gCDGunSightShader;  // @0x10DE4D0
